centralize cleanup jobs
